Refactored network classes
