+++ src/transport/network.cpp
@@ -208,265 +208,6 @@ void init(const M_string& remappings)
   }
 }
 
-NetAddress::NetAddress()
-{
-}
-
-NetAddress::NetAddress(NetAddress&& other) noexcept
-{
-  std::swap(rawAddress, other.rawAddress);
-  std::swap(rawAddressSize, other.rawAddressSize);
-  std::swap(address, other.address);
-
-  port = other.port;
-  type = other.type;
-}
-
-NetAddress::NetAddress(const NetAddress& other)
-{
-  if (other.rawAddress && (other.type == Type::AddressIPv4 || other.type == Type::AddressIPv6)) {
-    sockaddr_in* addr = static_cast<sockaddr_in*>(malloc(sizeof(sockaddr_in)));
-    rawAddress = addr;
-    rawAddressSize = other.rawAddressSize;
-    memcpy(rawAddress, other.rawAddress, other.rawAddressSize);
-    type = other.type;
-  } else {
-    type = Type::AddressInvalid;
-  }
-  port = other.port;
-  address = other.address;
-}
-
-NetAddress::~NetAddress()
-{
-  reset();
-}
-
-NetAddress& NetAddress::operator=(const NetAddress& other)
-{
-  if (this == &other)
-    return *this;
-  reset();
-
-  if (other.rawAddress && (other.type == Type::AddressIPv4 || other.type == Type::AddressIPv6)) {
-    sockaddr_in* addr = static_cast<sockaddr_in*>(malloc(sizeof(sockaddr_in)));
-    rawAddress = addr;
-    rawAddressSize = other.rawAddressSize;
-    memcpy(rawAddress, other.rawAddress, other.rawAddressSize);
-    type = other.type;
-  } else {
-    type = Type::AddressInvalid;
-  }
-  port = other.port;
-  address = other.address;
-  return *this;
-}
-
-void NetAddress::reset()
-{
-  if (rawAddress) {
-    free(rawAddress);
-    rawAddress = nullptr;
-  }
-  address = "";
-  port = 0;
-  type = Type::AddressInvalid;
-}
-
-bool NetAddress::isLocal() const
-{
-  return address == "127.0.0.1" || address == "localhost";
-}
-
-bool fillAddress(const sockaddr_in& my_addr, int len, NetAddress& address)
-{
-  address.reset();
-
-  if (my_addr.sin_family == AF_INET) {
-    address.type = NetAddress::AddressIPv4;
-    address.port = ntohs(my_addr.sin_port);
-  } else if (my_addr.sin_family == AF_INET6) {
-    address.type = NetAddress::AddressIPv6;
-    address.port = ntohs(my_addr.sin_port);
-  } else {
-    address.type = NetAddress::AddressInvalid;
-    return false;
-  }
-
-  char ipBuffer[255];
-  if (!inet_ntop(my_addr.sin_family, &my_addr.sin_addr, ipBuffer, sizeof(ipBuffer))) {
-    address.port = 0;
-    return false;
-  }
-  address.address = ipBuffer;
-
-  sockaddr_in* outAddr = static_cast<sockaddr_in*>(malloc(sizeof(sockaddr_in)));
-  address.rawAddress = outAddr;
-  address.rawAddressSize = len;
-  memcpy(outAddr, &my_addr, sizeof(sockaddr_in));
-  return true;
-}
-
-/// Fills in local address from socket.
-bool readLocalAddress(int sockfd, NetAddress& address)
-{
-  sockaddr_in my_addr{};
-  socklen_t len = sizeof(my_addr);
-  if (getsockname(sockfd, (sockaddr*)&my_addr, &len) != 0)
-    return false;
-
-  return fillAddress(my_addr, len, address);
-}
-
-/// Fills in remote address from socket.
-bool readRemoteAddress(int sockfd, NetAddress& address)
-{
-  sockaddr_in my_addr{};
-  socklen_t len = sizeof(my_addr);
-  if (getpeername(sockfd, (sockaddr*)&my_addr, &len) != 0)
-    return false;
-
-  return fillAddress(my_addr, len, address);
-}
-
-URL::URL()
-{
-}
-
-bool URL::fromString(const std::string& uri, bool defaultPort)
-{
-  reset();
-
-  // http://192.156.54.23/:11223/RPC2/request?.......
-
-  constexpr char schemeMarker[] = "://";
-  std::string::size_type schemePos = uri.find_first_of(schemeMarker);
-  if (schemePos == std::string::npos)
-    return false;
-
-  schemePos += sizeof(schemeMarker) - 1;
-  scheme = uri.substr(0, schemePos);
-
-  std::string::size_type partPos = uri.find_first_of('/', schemePos);
-
-  // Port is somewhere between hostname and resource part.
-  std::string::size_type portPos = uri.find_first_of(':', schemePos);
-  if (portPos == std::string::npos) {
-
-  } else {
-
-  }
-
-  host = uri.substr(schemePos);
-
-  // split out the port
-  std::string::size_type colon_pos = host.find_first_of(":");
-  if (colon_pos == std::string::npos)
-    return false;
-  std::string port_str = host.substr(colon_pos + 1);
-  std::string::size_type slash_pos = port_str.find_first_of("/");
-  if (slash_pos != std::string::npos)
-    port_str = port_str.erase(slash_pos);
-  port = atoi(port_str.c_str());
-  host = host.erase(colon_pos);
-  return true;
-}
-
-void URL::reset()
-{
-  host = {};
-  port = 0;
-  path = {};
-  scheme = {};
-}
-
-bool URL::empty() const
-{
-  return host.empty();
-}
-
-std::string URL::toString() const
-{
-  std::stringstream ss;
-  ss << scheme << host;
-  if (port)
-    ss << ":" << port;
-  if (path.empty())
-    ss << "/" << path;
-  return ss.str();
-}
-
-const char* addressFamilyName(sa_family_t family)
-{
-  switch (family) {
-    case AF_INET: return "AF_INET";
-    case AF_INET6: return "AF_INET6";
-    case AF_PACKET: return "AF_PACKET";
-    default:
-      return "????";
-  }
-}
-
-int listNetworkInterfaces()
-{
-#ifdef HAVE_IFADDRS_H
-  ifaddrs* addressList = nullptr;
-
-  int ret = getifaddrs(&addressList);
-  if (ret < 0) {
-    return ret;
-  }
-  if (addressList == nullptr)
-    return -1;
-
-  for (const ifaddrs *ifa = addressList; ifa; ifa = ifa->ifa_next) {
-    if (!ifa->ifa_addr)
-      continue;
-
-    sa_family_t family = ifa->ifa_addr->sa_family;
-    if (family == AF_INET || family == AF_INET6) {
-      /* Display interface name and family (including symbolic form of the latter for the common families). */
-      printf("%-8s %s (%d)\n", ifa->ifa_name, addressFamilyName(family), family);
-
-      char tmpHost[NI_MAXHOST];
-      /* For an AF_INET* interface address, display the address. */
-      int s = getnameinfo(ifa->ifa_addr, (family == AF_INET) ? sizeof(struct sockaddr_in) :
-                                    sizeof(struct sockaddr_in6),
-              tmpHost, NI_MAXHOST, NULL, 0, NI_NUMERICHOST);
-      if (s != 0) {
-        printf("getnameinfo() failed: %s\n", gai_strerror(s));
-      }
-
-      std::string host = tmpHost;
-
-      /* For an AF_INET* interface address, display the address. */
-      s = getnameinfo(ifa->ifa_netmask, (family == AF_INET) ? sizeof(struct sockaddr_in) :
-                                    sizeof(struct sockaddr_in6),
-              tmpHost, NI_MAXHOST, NULL, 0, NI_NUMERICHOST);
-      if (s != 0) {
-        printf("getnameinfo() failed: %s\n", gai_strerror(s));
-      }
-      std::string mask = tmpHost;
-      printf("\t\taddress: <%s> mask: <%s>\n", host.c_str(), mask.c_str());
-    }
-
-    /*
-    else if (family == AF_PACKET && ifa->ifa_data != NULL) {
-      struct rtnl_link_stats *stats = ifa->ifa_data;
-
-      printf("\t\ttx_packets = %10u; rx_packets = %10u\n"
-             "\t\ttx_bytes   = %10u; rx_bytes   = %10u\n",
-             stats->tx_packets, stats->rx_packets,
-             stats->tx_bytes, stats->rx_bytes);
-    }*/
-  }
-
-  if (addressList)
-    freeifaddrs(addressList);
-#endif
-  return 0;
-}
-
 } // namespace network
 
 } // namespace miniros
